cli: move find_config to sopel.cli.utils
