+++ sopel/cli/__init__.py
@@ -2,4 +2,4 @@
 from __future__ import unicode_literals, absolute_import, print_function, division
 
 # Shortcut imports
-from .utils import enumerate_configs  # noqa
+from .utils import enumerate_configs, find_config  # noqa
