+++ test/test_run_script.py
@@ -32,43 +32,6 @@ def config_dir(tmpdir):
     return test_dir
 
 
-def test_find_config_local(tmpdir, config_dir):
-    """Assert function retrieves configuration file from working dir first"""
-    working_dir = tmpdir.mkdir("working")
-    working_dir.join('local.cfg').write('')
-
-    with cd(working_dir.strpath):
-        found_config = run_script.find_config(config_dir.strpath, 'local.cfg')
-        assert found_config == 'local.cfg'
-
-        found_config = run_script.find_config(config_dir.strpath, 'local')
-        assert found_config == config_dir.join('local').strpath
-
-
-def test_find_config_default(tmpdir, config_dir):
-    """Assert function retrieves configuration file from given config dir"""
-    working_dir = tmpdir.mkdir("working")
-    working_dir.join('local.cfg').write('')
-
-    with cd(working_dir.strpath):
-        found_config = run_script.find_config(config_dir.strpath, 'config')
-        assert found_config == config_dir.join('config.cfg').strpath
-
-        found_config = run_script.find_config(config_dir.strpath, 'config.cfg')
-        assert found_config == config_dir.join('config.cfg').strpath
-
-
-def test_find_config_extension(tmpdir, config_dir):
-    """Assert function retrieves configuration file with the given extension"""
-    working_dir = tmpdir.mkdir("working")
-    working_dir.join('local.cfg').write('')
-
-    with cd(working_dir.strpath):
-        found_config = run_script.find_config(
-            config_dir.strpath, 'extra', '.ini')
-        assert found_config == config_dir.join('extra.ini').strpath
-
-
 def test_get_configuration(tmpdir):
     """Assert function returns a Sopel ``Config`` object"""
     working_dir = tmpdir.mkdir("working")
